+++ tests/test_spans.py
@@ -12,35 +12,35 @@ class TestSpans(unittest.TestCase):
         span = (0, 4, 1)
         expected_output = [(0, 1, 0), (0, 2, 0), (0, 3, 0), (1, 2, 0), (1, 3, 0), (1, 4, 0), (2, 3, 0),
                            (2, 4, 0), (3, 4, 0)]
-        self.assertEqual(build_non_enzymatic_spans(span), expected_output)
+        self.assertEqual(list(build_non_enzymatic_spans(span)), expected_output)
 
         span = (0, 3, 2)
         min_len = 1
         max_len = 2
         expected_output = [(0, 1, 0), (0, 2, 0), (1, 2, 0), (1, 3, 0), (2, 3, 0)]
-        self.assertEqual(build_non_enzymatic_spans(span, min_len, max_len), expected_output)
+        self.assertEqual(list(build_non_enzymatic_spans(span, min_len, max_len)), expected_output)
 
     def test_get_left_spans(self):
         span = (0, 4, 1)
         expected_output = [(0, 3, 1), (0, 2, 1), (0, 1, 1)]
-        self.assertEqual(build_left_semi_spans(span), expected_output)
+        self.assertEqual(list(build_left_semi_spans(span)), expected_output)
 
         span = (0, 3, 2)
         min_len = 1
         max_len = 2
         expected_output = [(0, 2, 2), (0, 1, 2)]
-        self.assertEqual(build_left_semi_spans(span, min_len, max_len), expected_output)
+        self.assertEqual(list(build_left_semi_spans(span, min_len, max_len)), expected_output)
 
     def test_get_right_spans(self):
         span = (0, 4, 1)
         expected_output = [(1, 4, 1), (2, 4, 1), (3, 4, 1)]
-        self.assertEqual(build_right_semi_spans(span), expected_output)
+        self.assertEqual(list(build_right_semi_spans(span)), expected_output)
 
         span = (0, 3, 2)
         min_len = 1
         max_len = 2
         expected_output = [(1, 3, 2), (2, 3, 2)]
-        self.assertEqual(build_right_semi_spans(span, min_len, max_len), expected_output)
+        self.assertEqual(list(build_right_semi_spans(span, min_len, max_len)), expected_output)
 
     def testspan_to_sequence(self):
         sequence = "ABCDEFGH"
@@ -56,17 +56,17 @@ class TestSpans(unittest.TestCase):
     def test_get_non_enzymatic_spans_no_sub_spans(self):
         span = (0, 0, 1)
         expected_output = []
-        self.assertEqual(build_non_enzymatic_spans(span), expected_output)
+        self.assertEqual(list(build_non_enzymatic_spans(span)), expected_output)
 
     def test_get_left_spans_span_of_one(self):
         span = (0, 0, 1)
         expected_output = []
-        self.assertEqual(build_left_semi_spans(span), expected_output)
+        self.assertEqual(list(build_left_semi_spans(span)), expected_output)
 
     def test_get_right_spans_span_of_one(self):
         span = (0, 0, 1)
         expected_output = []
-        self.assertEqual(build_right_semi_spans(span), expected_output)
+        self.assertEqual(list(build_right_semi_spans(span)), expected_output)
 
     def testspan_to_sequence_empty_sequence(self):
         sequence = ""
@@ -93,17 +93,17 @@ class TestSpans(unittest.TestCase):
     def test_get_non_enzymatic_spans_non_integer_input(self):
         span = (0.5, 4.5, 1)
         with self.assertRaises(TypeError):
-            build_non_enzymatic_spans(span)
+            list(build_non_enzymatic_spans(span))
 
     def test_get_left_spans_non_integer_input(self):
         span = (0.5, 4.5, 1)
         with self.assertRaises(TypeError):
-            build_left_semi_spans(span)
+            list(build_left_semi_spans(span))
 
     def test_get_right_spans_non_integer_input(self):
         span = (0.5, 4.5, 1)
         with self.assertRaises(TypeError):
-            build_right_semi_spans(span)
+            list(build_right_semi_spans(span))
 
     def testspan_to_sequence_non_integer_input(self):
         sequence = "ABCDEFGH"
@@ -119,7 +119,7 @@ class TestSpans(unittest.TestCase):
         max_len = 6
 
         expected_spans = [(0, 2, 0), (0, 5, 1), (2, 5, 0), (2, 8, 1), (5, 8, 0), (5, 10, 1), (8, 10, 0)]
-        spans = build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len)
+        spans = list(build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len))
         self.assertEqual(spans, expected_spans)
 
     def test_get_enzymatic_spans2(self):
@@ -130,7 +130,7 @@ class TestSpans(unittest.TestCase):
         max_len = 5
 
         expected_spans = [(0, 2, 0), (0, 5, 1), (2, 5, 0), (5, 8, 0), (5, 10, 1), (8, 10, 0)]
-        spans = build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len)
+        spans = list(build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len))
         self.assertEqual(spans, expected_spans)
 
     def test_get_enzymatic_spans_oob(self):
@@ -142,7 +142,7 @@ class TestSpans(unittest.TestCase):
 
         expected_spans = [(0, 2, 0), (0, 5, 1), (0, 8, 2), (0, 10, 3), (2, 5, 0), (2, 8, 1), (2, 10, 2), (5, 8, 0),
                           (5, 10, 1), (8, 10, 0)]
-        spans = build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len)
+        spans = list(build_enzymatic_spans(end_index, enzyme_sites, missed_cleavages, min_len, max_len))
         self.assertEqual(spans, expected_spans)
 
     def test_get_semi_spans(self):
@@ -152,7 +152,7 @@ class TestSpans(unittest.TestCase):
 
         expected_semi_spans = [(0, 4, 0), (0, 3, 0), (0, 2, 0), (1, 5, 0), (2, 5, 0), (3, 5, 0), (3, 8, 1),
                                (4, 8, 1), (5, 8, 1), (6, 8, 1)]
-        semi_spans = build_semi_spans(spans, min_len, max_len)
+        semi_spans = list(build_semi_spans(spans, min_len, max_len))
         self.assertEqual(semi_spans, expected_semi_spans)
 
     def test_get_semi_spans2(self):
@@ -163,7 +163,7 @@ class TestSpans(unittest.TestCase):
         expected_semi_spans = [(10, 14, 1), (10, 13, 1), (10, 12, 1), (11, 15, 1), (12, 15, 1), (13, 15, 1),
                                (13, 18, 2),
                                (14, 18, 2), (15, 18, 2), (16, 18, 2)]
-        semi_spans = build_semi_spans(spans, min_len, max_len)
+        semi_spans = list(build_semi_spans(spans, min_len, max_len))
         self.assertEqual(semi_spans, expected_semi_spans)
 
     def test_get_semi_spans3(self):
@@ -173,7 +173,7 @@ class TestSpans(unittest.TestCase):
 
         expected_semi_spans = [(5, 10, 2), (5, 9, 2), (5, 8, 2), (5, 7, 2), (10, 14, 1), (10, 13, 1), (10, 12, 1),
                                (11, 15, 1), (12, 15, 1), (13, 15, 1)]
-        semi_spans = build_semi_spans(spans, min_len, max_len)
+        semi_spans = list(build_semi_spans(spans, min_len, max_len))
         self.assertEqual(semi_spans, expected_semi_spans)